feat: Add course duplication functionality
Introduces the ability to duplicate existing courses within the study portal. This will streamline the process for instructors to create new courses based on existing ones, saving time and effort.

The `duplicateCourse` function has been added to the store, and the UI will be updated in subsequent commits to expose this feature.